README: full Y2 scope; clippy cleanup across the workspace
README now documents the whole engine — all 12 crates, plus sections for
aggregations/analytics, keyword+vector+hybrid ranking, the live NLQ loop
(OpenAI-compatible/Groq/Bedrock), and the agent-platform layer (governed
memory, federation, SIEM generality, capstone). Clippy: applied the safe
idiomatic fixes workspace-wide (is_none_or/is_some_and, redundant closures,
Option::replace, is_multiple_of), annotated the one intentional
await-holding-lock (the backpressure test), and de-duplicated a doc comment.
Remaining warnings are cosmetic (doc list formatting). Unit tests green.

+++ crates/coordinator/tests/aggregate_correctness.rs
@@ -18,7 +18,7 @@ fn corpus(n: usize, seed: u64) -> Vec<FlightDocument> {
     let mut rng = seed;
     let mut next = move || {
         rng = rng.wrapping_mul(6364136223846793005).wrapping_add(1442695040888963407);
-        (rng >> 33) as u64
+        (rng >> 33)
     };
     let origins = ["SFO", "JFK", "ATL", "ORD", "DEN", "LAX"];
     let crafts = ["Boeing 737", "Airbus A320", "Boeing 777", "Embraer E190"];
